updated everythign & made it all TSX
