+++ AutoDungeoners.Web/ClientApp/package.json
@@ -3,47 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.7",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.5.0",
+    "@testing-library/user-event": "^7.2.1",
+    "@types/jest": "^24.9.1",
+    "@types/jwt-decode": "^2.2.1",
+    "@types/node": "^12.12.41",
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-router": "^5.1.7",
     "@types/react-router-dom": "^5.1.5",
     "@types/reactstrap": "^8.4.2",
-    "babel-eslint": "10.0.1",
-    "bootstrap": "^4.1.3",
-    "jquery": "^3.5.0",
+    "bootstrap": "^4.5.0",
     "jwt-decode": "^2.2.0",
-    "merge": "^1.2.1",
-    "oidc-client": "^1.9.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-router-bootstrap": "^0.24.4",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "^3.0.1",
-    "reactstrap": "^6.3.0",
-    "rimraf": "^2.6.2"
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.1",
+    "reactstrap": "^8.4.1",
+    "typescript": "^3.7.5"
   },
-  "devDependencies": {
-    "ajv": "^6.9.1",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.12.0",
-    "eslint-config-react-app": "^4.0.1",
-    "eslint-plugin-flowtype": "^2.0.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.11.1",
-    "source-map-loader": "^0.2.4",
-    "ts-loader": "^7.0.3",
-    "typescript": "^3.8.3",
-    "webpack-cli": "^3.3.11"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "scripts": {
-    "start": "rimraf ./build && react-scripts start",
-    "build": "react-scripts build",
-    "test": "cross-env CI=true react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "lint": "eslint ./src/"
   },
   "browserslist": {
     "production": [
