+++ AutoDungeoners.Web/ClientApp/src/components/Authentication/Identity.test.js
