+++ AutoDungeoners.Web/ClientApp/src/components/Authentication/Identity.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { NavLink } from 'reactstrap';
 import { Link } from 'react-router-dom';
+import JwtBody from '../../interfaces/Jwt';
 import jwtDecode from 'jwt-decode';
 
 interface TokenProps {
@@ -9,6 +10,6 @@ interface TokenProps {
 
 export function Identity(props:TokenProps) {
   return (
-    <NavLink tag={Link} className="text-dark" to="/">{ props.token != null ? "Hi, " + jwtDecode(props.token).email : "Unauthenticated" }</NavLink>
+    <NavLink tag={Link} className="text-dark" to="/">{ props.token != null ? "Hi, " + (jwtDecode(props.token) as JwtBody).email : "Unauthenticated" }</NavLink>
   )
 }
