+++ AutoDungeoners.Web/ClientApp/src/index.tsx
@@ -1,0 +1,22 @@
+import 'bootstrap/dist/css/bootstrap.css';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+import { BrowserRouter } from 'react-router-dom';
+import * as serviceWorker from './serviceWorker';
+
+const baseUrl = document.getElementsByTagName('base')[0].getAttribute('href');
+
+ReactDOM.render(
+  <React.StrictMode>
+    <BrowserRouter basename={baseUrl!}>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>,
+  document.getElementById('root')
+);
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
