+++ AutoDungeoners.Web/ClientApp/src/interfaces/Jwt.tsx
@@ -1,0 +1,6 @@
+
+export default interface JwtBody {
+    sub: string,
+    email: string,
+    jti: Number
+}
