+++ AutoDungeoners.Web/ClientApp/tsconfig.json
@@ -1,30 +1,26 @@
 {
   "compilerOptions": {
-    "noEmitOnError": true,
-    "outDir": "./dist/",
-    "sourceMap": true,
-    "strictNullChecks": false,
-    "module": "esnext",
-    "jsx": "preserve",
     "target": "es5",
-    "allowJs": true,
     "lib": [
       "dom",
       "dom.iterable",
       "esnext"
     ],
+    "allowJs": false,
     "skipLibCheck": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
+    "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "noImplicitAny": false
+    "jsx": "react",
+    "strictNullChecks": true,
   },
   "include": [
-    "./src/"
+    "src"
   ]
 }
